% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{data_calMean_calSD}
\alias{data_calMean_calSD}
\title{Calculate mean and sd value by conditions}
\usage{
data_calMean_calSD(prot_dat, meta)
}
\arguments{
\item{prot_dat}{A DataFrame from previous processing.}

\item{meta}{A DataFrame of meta file.}
}
\value{
A DataFrame of mean and sd by conditions.
}
\description{
Calculate mean and sd value by conditions.
}
\details{
Input dataframe from previous processing, then return a data set with mean and sd values by conditions.
}
