% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{data_draw_specific}
\alias{data_draw_specific}
\title{Draw Specific Protein Intensity plot}
\usage{
data_draw_specific(prot_dat_specific, id)
}
\arguments{
\item{prot_dat_specific}{A DataFrame from data_transform_melt.}

\item{id}{A integer or string.}
}
\value{
A ggplot object.
}
\description{
Draw Specific Protein Intensity plot.
}
\details{
Input dataframe from data_transform_melt, then return a ggplot object of intensity plot.
}
