% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean_normalization.R
\name{data_norm_ers}
\alias{data_norm_ers}
\title{Normalize MS TMT data with ERS.}
\usage{
data_norm_ers(prot_dat, condition_num, repeat_num, ref_col = c(1, 7, 13))
}
\arguments{
\item{prot_dat}{A DataFrame from read_maxquant_prot() with only numeric value.}
}
\value{
A DataFrame.
}
\description{
Normalize MS TMT data with External Reference Signal.
}
\details{
Input dataframe from read_maxquant_prot() function, then return a data set after normalization with ERS.
}
