% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean_normalization.R
\name{data_scale_mat}
\alias{data_scale_mat}
\title{Dataset Scaled.}
\usage{
data_scale_mat(prot)
}
\arguments{
\item{prot}{A DataFrame from with only numeric value.}
}
\description{
Scale dataframe into mean 0 and sd 1 by rows.
}
\details{
Input dataframe, then return a scaled set of dataframe.
}
