% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{data_transform_melt}
\alias{data_transform_melt}
\title{Transform DataFrame by melt}
\usage{
data_transform_melt(prot_dat, meta)
}
\arguments{
\item{prot_dat}{A DataFrame from previous processing.}

\item{meta}{A DataFrame of meta file.}
}
\value{
A DataFrame of melt.
}
\description{
Transform DataFrame by melt.
}
\details{
Input dataframe from previous processing, then return a DataFame of conditions melt.
}
