% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean_normalization.R
\name{data_zero_substitution_3repeats}
\alias{data_zero_substitution_3repeats}
\title{Zero Substitution in 3 Repeats.}
\usage{
data_zero_substitution_3repeats(
  protdat,
  s1,
  e1,
  s2,
  e2,
  s3,
  e3,
  condition_num,
  least = 30
)
}
\arguments{
\item{protdat}{A DataFrame from with only numeric value.}

\item{s1}{A integer of start index of repeat 1.}

\item{e1}{A integer of end index of repeat 1.}

\item{s2}{A integer of start index of repeat 2.}

\item{e2}{A integer of end index of repeat 2.}

\item{s3}{A integer of start index of repeat 3 need substitution.}

\item{e3}{A integer of end index of repeat 3 need substitution.}

\item{condition_num}{A integer of condition numbers.}

\item{least}{A integer for zero substitution for repeat 1 and 2.}
}
\description{
Zero value substitution of 1 repeat in 3 repeats with average.
}
\details{
Input dataframe, and column start and end index of each repeat, then return a set of dataframe.
}
