% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read.R
\docType{class}
\name{MSDataSet-class}
\alias{MSDataSet-class}
\alias{MSDataSet}
\title{A Reference Class to represent MS dataset.}
\description{
A Reference Class to represent MS dataset.
}
\section{Fields}{

\describe{
\item{\code{balance}}{A length-one numeric vector.}
}}


