% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read.R
\name{read_maxquant_prot}
\alias{read_maxquant_prot}
\title{Read, clean and annotate MS TMT data.}
\usage{
read_maxquant_prot(
  prot_raw,
  meta,
  name = "defalut",
  reference = FALSE,
  analysis_type = "TMT",
  exclusion_status = "none",
  normalization_status = c("none")
)
}
\arguments{
\item{prot_raw}{A DataFrame of proteinGroups from MaxQuant.}

\item{meta}{A DataFrame of metafile.}

\item{name}{A character of the name.}

\item{reference}{A boolean.}

\item{analysis_type}{A character.}

\item{exclusion_status}{A character.}

\item{normalization_status}{A vector.}
}
\description{
Read, clean and annotate MS TMT data with Maxquant process data and meta file.
}
\details{
Input MaxQuang processed MS TMT data and meta file, then return a MSDataSet object.
}
